requirements/requirements.txt
chromadb
ollama
python-dotenv
tavily
elevenlabs
silero
onnxruntime 
soundfile 
numpy
audioop-lts
yt-dlp 
sounddevice
flask
flask_cors
-e ./Visualizer/point-e
faster-whisper
huggingface_hub
PyPDF2
pytesseract
python-docx
pandas
openpyxl
pdf2image
pyinstaller
beautifulsoup4
detoxify 
transformers 
rapidfuzz
pywebview
waitress
pythonnet